Move DatePicker logic into DatePickerBase<T> with abstract conversion methods
- Add abstract methods ToDateOnly(T), FromDateOnly(DateOnly), Parse(string, string), Format(T, string)
- Add virtual GetToday() returning DateOnly.FromDateTime(DateTime.Today)
- Move all template parts, event handlers, IClearControl, and property change logic to base
- DatePicker now only implements the 4 abstract methods and holds placeholder/watermark properties
